refactor: merge session date helper + de-dup agent transcript save (wave 3.5)
- §2.8 move sessionUtils.formatDate → lib/utils formatDateCompact; delete sessionUtils.ts
- §3.5 extract persistPiTranscript() shared by AgentChatPane onDone/onError
- add changelogs/v2.3.6.md; log wave 3.5 in architecture-review.md
