+++ changelogs/v2.3.6.md
@@ -1,0 +1,16 @@
+## What's new
+
+A maintenance release focused on internal cleanup: shared hooks and components were extracted across the app to remove duplicated logic and shrink several oversized files. No new features, but it fixes a few UI consistency issues along the way.
+
+### Fixes
+
+- **Code, diff, and git views now recolour instantly on theme change**: Syntax-highlighted code blocks, the diff viewer, and the git panel read the theme once when they rendered and didn't update when you switched light/dark (or when the OS theme changed in "system" mode) — leaving stale colours until the view remounted. They now repaint immediately on a theme change.
+- **More reliable navigation from links and chips**: Opening a note or task from a card's linked-notes list, a chat/agent reference chip, a flow node, the search panel, or the project overview occasionally raced the target view's load and failed to scroll to / select the item. These now consistently switch views and reveal the target.
+- **Consistent syntax-highlight colours everywhere**: Code blocks in notes, the dashboard HTML editor, and PDF export all draw from a single shared palette, so a given token renders the same colour across every surface (and the PDF light-theme remap stays in sync).
+
+### Changes
+
+- **Medium-priority colour aligned to blue**: The medium priority level now renders blue (`--info`) consistently. Previously the class-based badges (flow task nodes, the node-edit modal, the top-bar project badge, the card-detail flag) showed amber while the board card stripes and analytics charts already used blue. They now all match.
+- **Shared UI primitives**: Introduced reusable building blocks — a theme hook (`useIsDark`), a copy-to-clipboard hook, a navigation helper, a `Toggle` switch, and a node-type chip — and adopted them across chat, agent, notes, kanban, graph, insights, settings, and onboarding to replace ad-hoc copies.
+- **Unified the card detail view**: The card-detail modal and the inline card-detail panel (shown beside chat) now share one implementation, so they always stay in sync.
+- **Smaller, more focused modules**: The kanban board's archive view, the note PDF-export logic, and the shared markdown code-block renderer were split into their own files; the store's startup hydration paths now share a single restore routine. The agent chat pane's turn-complete and error handlers now share one transcript-save routine, and the session-list date formatter moved into the shared date helpers.
