+++ src/components/agent/AgentChatPane.tsx
@@ -60,6 +60,26 @@ function extractCairnRef(
   }
 }
 
+/**
+ * Persist the session's finalised message transcript to the backend. Reads the
+ * latest store state imperatively (so it is safe to call from IPC callbacks),
+ * drops still-streaming messages, and fire-and-forgets the save. Shared by the
+ * `onDone` and `onError` handlers, which previously inlined identical blocks.
+ */
+function persistPiTranscript(sessionId: string): void {
+  const msgs = useCairnStore.getState().terminalSessions.find((t) => t.sessionId === sessionId)?.piMessages ?? [];
+  const saveable = msgs.filter((m) => !m.isStreaming).map((m) => ({
+    id: m.id,
+    role: m.role,
+    content: m.content,
+    reasoning: m.reasoning ?? null,
+    toolCalls: m.toolCalls ?? null,
+    subagents: m.subagents ?? null,
+    timestamp: m.timestamp,
+  }));
+  window.electron?.piAgent.saveMessages(sessionId, saveable).catch(console.error);
+}
+
 
 const NATIVE_AGENT_SLASH_COMMANDS = [
   {
@@ -248,17 +268,7 @@ export function AgentChatPane({ session, isActive }: AgentChatPaneProps) {
       setRetryInfo(null);
       setIsCompacting(false);
       // Persist the full message transcript after the turn completes
-      const msgs = useCairnStore.getState().terminalSessions.find((t) => t.sessionId === sessionId)?.piMessages ?? [];
-      const saveable = msgs.filter((m) => !m.isStreaming).map((m) => ({
-        id: m.id,
-        role: m.role,
-        content: m.content,
-        reasoning: m.reasoning ?? null,
-        toolCalls: m.toolCalls ?? null,
-        subagents: m.subagents ?? null,
-        timestamp: m.timestamp,
-      }));
-      window.electron?.piAgent.saveMessages(sessionId, saveable).catch(console.error);
+      persistPiTranscript(sessionId);
     });
 
     const unsubError = electron.piAgent.onError((e) => {
@@ -274,19 +284,7 @@ export function AgentChatPane({ session, isActive }: AgentChatPaneProps) {
       });
       setIsLoading(false);
       // Persist the message transcript including the error message
-      setTimeout(() => {
-        const msgs = useCairnStore.getState().terminalSessions.find((t) => t.sessionId === sessionId)?.piMessages ?? [];
-        const saveable = msgs.filter((m) => !m.isStreaming).map((m) => ({
-          id: m.id,
-          role: m.role,
-          content: m.content,
-          reasoning: m.reasoning ?? null,
-          toolCalls: m.toolCalls ?? null,
-          subagents: m.subagents ?? null,
-          timestamp: m.timestamp,
-        }));
-        window.electron?.piAgent.saveMessages(sessionId, saveable).catch(console.error);
-      }, 0);
+      setTimeout(() => persistPiTranscript(sessionId), 0);
     });
 
     // ── Subagent events (child session IDs routed back to parent) ──────────
