--- src/components/agent/sessionUtils.ts
@@ -1,20 +1,0 @@
-// Pure helper for formatting session timestamps.
-
-export function formatDate(iso: string) {
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) return "Invalid date";
-
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const target = new Date(d);
-  target.setHours(0, 0, 0, 0);
-
-  const diffDays = Math.floor((today.getTime() - target.getTime()) / 86400000);
-  if (diffDays < 0) {
-    return d.toLocaleDateString(undefined, { month: "short", day: "numeric" });
-  }
-  if (diffDays === 0) return "Today";
-  if (diffDays === 1) return "Yesterday";
-  if (diffDays < 7) return `${diffDays}d ago`;
-  return d.toLocaleDateString(undefined, { month: "short", day: "numeric" });
-}
